Improved motion

+++ code/tortoise.py
@@ -329,11 +329,11 @@ class Tortoise:
 
             if direction == enums.Direction.backward_left or direction == enums.Direction.backward or direction == enums.Direction.counterClockwise:
 
-                self.A.backwards(int(self.delay) / 1000.00, 1)
+                self.A.backward(int(self.delay) / 1000.00, 1)
 
             if direction == enums.Direction.backward_right or direction == enums.Direction.backward or direction == enums.Direction.clockwise:
 
-                self.B.backwards(int(self.delay) / 1000.00, 1)
+                self.B.backward(int(self.delay) / 1000.00, 1)
 
             if direction == enums.Direction.forward_right or direction == enums.Direction.forward or direction == enums.Direction.clockwise:
 
@@ -349,11 +349,11 @@ class Tortoise:
 
                     if direction == enums.Direction.backward_left or direction == enums.Direction.backward or direction == enums.Direction.counterClockwise:
 
-                        self.A.backwards(int(self.delay) / 1000.00, numberOfStepsRemaining)
+                        self.A.backward(int(self.delay) / 1000.00, numberOfStepsRemaining)
 
                     if direction == enums.Direction.backward_right or direction == enums.Direction.backward or direction == enums.Direction.clockwise:
 
-                        self.B.backwards(int(self.delay) / 1000.00, numberOfStepsRemaining)
+                        self.B.backward(int(self.delay) / 1000.00, numberOfStepsRemaining)
 
                     if direction == enums.Direction.forward_right or direction == enums.Direction.forward or direction == enums.Direction.clockwise:
 
@@ -412,38 +412,39 @@ class Tortoise:
                 self.setStateTortoise(enums.State.running)
                 print "[TORTOISE RESUMED]"
              
+            motorAprocess_backward = Process(target=self.A.backward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
+            motorBprocess_backward = Process(target=self.B.backward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
+            motorAprocess_forward = Process(target=self.A.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
+            motorBprocess_forward = Process(target=self.B.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
+
 
             if direction == enums.Direction.backward_left or direction == enums.Direction.backward or direction == enums.Direction.counterClockwise:
 
-                motorAprocess_backward = Process(target=self.A.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
                 motorAprocess_backward.start()
 
             if direction == enums.Direction.backward_right or direction == enums.Direction.backward or direction == enums.Direction.clockwise:
 
-                motorBprocess_backward = Process(target=self.B.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
                 motorBprocess_backward.start()
 
             if direction == enums.Direction.forward_right or direction == enums.Direction.forward or direction == enums.Direction.clockwise:
 
-                motorAprocess_forward = Process(target=self.A.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
                 motorAprocess_forward.start()
 
             if direction == enums.Direction.forward_left or direction == enums.Direction.forward or direction == enums.Direction.counterClockwise:
 
-                motorBprocess_forward = Process(target=self.B.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
                 motorBprocess_forward.start()
 
 
-            if motorAprocess_forward.is_alive()
+            if motorAprocess_forward.is_alive():
                 motorAprocess_forward.join()
 
-            if motorBprocess_forward.is_alive()
+            if motorBprocess_forward.is_alive():
                 motorBprocess_forward.join()
 
-            if motorAprocess_backward.is_alive()
+            if motorAprocess_backward.is_alive():
                 motorAprocess_backward.join()
 
-            if motorBprocess_backward.is_alive()
+            if motorBprocess_backward.is_alive():
                 motorBprocess_backward.join()
 
 
@@ -457,45 +458,19 @@ class Tortoise:
 
     def moveMotors_improved2(self, steps, direction):
 
-        if( direction != enums.Direction.backward_right and direction != enums.Direction.backward_left and 
+            if( direction != enums.Direction.backward_right and direction != enums.Direction.backward_left and 
             direction != enums.Direction.forward_right and direction != enums.Direction.forward_left and direction != enums.Direction.forward and direction != enums.Direction.backward ) :
-            print "Hey, my master! I can only move backward or forward, and either left or right."
-            print "\tHINT: check the direction ;)"
-            self.blinkLED(1, 3, 0.5)
-            return -1
 
-        if(steps < 0):
-            print "How am I going to move a negative number of steps? I can't travel back in time!"
-            print "\tHINT: check the number of steps ;)"
-            self.blinkLED(1, 3, 0.5)
-            return -1
+                    print "Hey, my master! I can only move backward or forward, and either left or right."
+                    print "\tHINT: check the direction ;)"
+                    self.blinkLED(1, 3, 0.5)
+                    return -1
 
-
-        motorAprocess_backward = Process(target=self.A.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
-        motorBprocess_backward = Process(target=self.B.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
-        motorAprocess_forward = Process(target=self.A.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
-        motorBprocess_forward = Process(target=self.B.forward, args=(int(self.delay) / 1000.00, numberOfstepsCommanded))
-
-
-        if direction == enums.Direction.backward_left or direction == enums.Direction.backward or direction == enums.Direction.counterClockwise:
-            
-            motorAprocess_backward.start()
-
-        if direction == enums.Direction.backward_right or direction == enums.Direction.backward or direction == enums.Direction.clockwise:
-
-            motorBprocess_backward.start()
-
-        if direction == enums.Direction.forward_right or direction == enums.Direction.forward or direction == enums.Direction.clockwise:
-
-            motorAprocess_forward.start()
-
-        if direction == enums.Direction.forward_left or direction == enums.Direction.forward or direction == enums.Direction.counterClockwise:
-
-            motorBprocess_forward.start()
-
-
-        # The main loop pools the emergencySwitch
-        while motorAprocess_backward.is_alive() or motorBprocess_backward.is_alive() or motorAprocess_forward.is_alive() or motorBprocess_forward.is_alive():
+            if(steps < 0):
+                    print "How am I going to move a negative number of steps? I can't travel back in time!"
+                    print "\tHINT: check the number of steps ;)"
+                    self.blinkLED(1, 3, 0.5)
+                    return -1
 
             # If a stop command has been sent, the turtle will stop its movement
             if self.getSensorData(enums.SensorType.emergencySwitch, 1) == 0:
@@ -505,34 +480,73 @@ class Tortoise:
                     self.setStateTortoise(enums.State.paused)
                     print "[TORTOISE PAUSED]"
 
-                    if motorAprocess_backward.is_alive():
-                        motorAprocess_backward.terminate()
-                        motorAprocess_backward.join()
+            else:
+                if self.getStateTortoise() == enums.State.paused:
+                        self.setStateTortoise(enums.State.running)
+                        print "[TORTOISE RESUMED]"
 
-                    if motorBprocess_backward.is_alive():
-                        motorBprocess_backward.terminate()
-                        motorBprocess_backward.join()
+                motorAprocess_backward = Process(target=self.A.backward, args=(int(self.delay) / 1000.00, steps))
+                motorBprocess_backward = Process(target=self.B.backward, args=(int(self.delay) / 1000.00, steps))
+                motorAprocess_forward = Process(target=self.A.forward, args=(int(self.delay) / 1000.00, steps))
+                motorBprocess_forward = Process(target=self.B.forward, args=(int(self.delay) / 1000.00, steps))
 
-                    if motorAprocess_forward.is_alive():
-                        motorAprocess_forward.terminate()
-                        motorAprocess_forward.join()
 
-                    if motorBprocess_forward.is_alive():
-                        motorBprocess_forward.terminate()
-                        motorBprocess_forward.join()
+                if direction == enums.Direction.backward_left or direction == enums.Direction.backward or direction == enums.Direction.counterClockwise:
                     
-            elif self.getStateTortoise() == enums.State.paused:
-                self.setStateTortoise(enums.State.running)
-                print "[TORTOISE RESUMED]"
+                    motorAprocess_backward.start()
+
+                if direction == enums.Direction.backward_right or direction == enums.Direction.backward or direction == enums.Direction.clockwise:
+
+                    motorBprocess_backward.start()
+
+                if direction == enums.Direction.forward_right or direction == enums.Direction.forward or direction == enums.Direction.clockwise:
+
+                    motorAprocess_forward.start()
+
+                if direction == enums.Direction.forward_left or direction == enums.Direction.forward or direction == enums.Direction.counterClockwise:
+
+                    motorBprocess_forward.start()
 
 
-            time.sleep(1)
+                # The main loop pools the emergencySwitch
+                while motorAprocess_backward.is_alive() or motorBprocess_backward.is_alive() or motorAprocess_forward.is_alive() or motorBprocess_forward.is_alive():
+
+                    # If a stop command has been sent, the turtle will stop its movement
+                    if self.getSensorData(enums.SensorType.emergencySwitch, 1) == 0:
+
+                        if self.getStateTortoise() == enums.State.running:
+
+                            self.setStateTortoise(enums.State.paused)
+                            print "[TORTOISE PAUSED]"
+
+                            if motorAprocess_backward.is_alive():
+                                motorAprocess_backward.terminate()
+                                motorAprocess_backward.join()
+
+                            if motorBprocess_backward.is_alive():
+                                motorBprocess_backward.terminate()
+                                motorBprocess_backward.join()
+
+                            if motorAprocess_forward.is_alive():
+                                motorAprocess_forward.terminate()
+                                motorAprocess_forward.join()
+
+                            if motorBprocess_forward.is_alive():
+                                motorBprocess_forward.terminate()
+                                motorBprocess_forward.join()
+                            
+                    elif self.getStateTortoise() == enums.State.paused:
+                        self.setStateTortoise(enums.State.running)
+                        print "[TORTOISE RESUMED]"
 
 
-        self.A.stopMotors()
-        self.B.stopMotors()
+                    time.sleep(0.5)
 
-        return 0
+
+            self.A.stopMotors()
+            self.B.stopMotors()
+
+            return 0
